Refactor code view into its own module
